+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 CHANGELOG
 =========
 
-2024-09-26 (0.13.1)
+2024-10-21 (0.13.2)
+------------------
+- Bugfix for link noise models depolarise and discrete depolarise in multithread variant
+- Added tests for link noise and hardware models for multithread variant
+
+2024-10-20 (0.13.1)
 ------------------
 - Bugfix for pydantic 2.x versions